requirements.txt
Django==4.2.1
psycopg2-binary==2.9.6
djangorestframework==3.14.0
pandas==1.5.2
pandera==0.13.4
